requirements.txt


Flask-Bootstrap~=3.3.7.1
Flask-Cors~=3.0.10
Flask-WTF~=0.15.1
Flask==2.0.1
WTForms~=2.3.3
Werkzeug~=2.0.1
aws-wsgi
awslambdaric==1.2.2
black==21.8b0
boto3~=1.18.36
botocore~=1.21.36
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
et-xmlfile==1.1.0
idna==3.2
mypy-extensions==0.4.3
numpy==1.21.2
openai==0.10.3
openpyxl==3.0.7
pandas-stubs==1.2.0.13
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
setuptools~=58.0.0
simplejson==3.17.2
six==1.16.0
tomli==1.2.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
visitor~=0.1.3
wheel~=0.37.0
